backend/requirements.txt
# /backend/requirements.txt

fastapi
uvicorn
openai
pydantic
python-dotenv
gunicorn
fastapi-cors